// src/app/shared/model/topology-data.ts
import { Graph } from "./graph";

export class TopologyData {
    graph?: Graph;
    topology?: string;
    orientation?: string;
    sequenceSubject?: any;
    sequenceDirection?: any;
    viewDisplayTitle?: string;
    viewShortTitle?: string;
    viewDisplayDesc?: string;
    objTitle?: string;
}